Automatically scroll to the newest message in chat
This commit adds functionality to automatically scroll to the bottom of the chat view when new messages arrive.

A `rememberLazyListState` is used to control the scroll position of the `LazyColumn` displaying chat messages.
A `LaunchedEffect` observes changes to the `items.size` (number of messages) and, if the list is not empty, calls `scrollToItem` to scroll to the last message.